공통 함수를 utils로 따로 정리 & 주석 정리
