--- launcher_abler/StateUI.py
@@ -1,10 +1,0 @@
-from enum import Enum, auto
-
-
-class StateUI(Enum):
-    check = auto()
-    no_release = auto()
-    update_launcher = auto()
-    update_abler = auto()
-    execute = auto()
-    error = auto()
